refactor: edit trigger component
