+++ src/components/newtab/workflow/edit/TriggerEvent/TriggerEventKeyboard.vue
