+++ src/components/newtab/workflow/edit/TriggerEvent/TriggerEventMouse.vue
