+++ src/components/newtab/workflow/edit/TriggerEvent/TriggerEventWheel.vue
