+++ src/manifest.chrome.json
@@ -47,7 +47,8 @@
   ],
   "optional_permissions": [
     "clipboardRead",
-    "downloads"
+    "downloads",
+    "contextMenus"
   ],
   "permissions": [
     "tabs",
